% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_forecast.R
\name{send_forecast}
\alias{send_forecast}
\title{Function to send a single forecast}
\usage{
send_forecast(group, forecast, key, date = paste(Sys.Date() + 1))
}
\arguments{
\item{group}{Group number (integer)}

\item{forecast}{Single forecast (numeric)}

\item{key}{The secret key corresponding to the group number}

\item{date}{The date ('YYYY-MM-DD' format). Default is the date of the system plus one day.}
}
\description{
Function to send a single forecast
}
